Moved siren to Wink from switch
